sorted api routes
tried to get rid of getStaticProps functions
