+++ tsconfig.json
@@ -20,6 +20,6 @@
       "@/*": [ "./*" ]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "yandex-functions/populateDB/populateDB.js", "pages/api/auth.js", "pages/api/auth/[...nextauth].js", "pages/protectedPage.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "yandex-functions/populateDB/populateDB.js", "pages/api/auth.js", "pages/api/auth/[...nextauth].js", "pages/protectedPage.js", "components/fetchParticipants.js"],
   "exclude": ["node_modules"]
 }
